# Remove LoadingCache and replace with array