--- a/src/main/java/shortestpath/pathfinder/SplitFlagMap.java
+++ b/src/main/java/shortestpath/pathfinder/SplitFlagMap.java
@@ -1,56 +1,59 @@
 package shortestpath.pathfinder;
 
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
-import com.google.common.cache.Weigher;
-import com.google.common.util.concurrent.UncheckedExecutionException;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
 import java.util.zip.GZIPInputStream;
 import shortestpath.Util;
 
 import static net.runelite.api.Constants.REGION_SIZE;
 
 public abstract class SplitFlagMap {
-    private static final int MAXIMUM_SIZE = 20 * 1024 * 1024;
 
-    private final LoadingCache<Integer, FlagMap> regionMaps;
+    // Size is automatically chosen based on the max extents of the collision data
+    public final FlagMap[] regionMaps;
+    private final CollisionMap.RegionExtent regionExtents;
+    private final int widthInclusive;
     private final int flagCount;
 
     public SplitFlagMap(Map<Integer, byte[]> compressedRegions, int flagCount) {
         this.flagCount = flagCount;
-        regionMaps = CacheBuilder
-                .newBuilder()
-                .weigher((Weigher<Integer, FlagMap>) (k, v) -> v.flags.size() / 8)
-                .maximumWeight(MAXIMUM_SIZE)
-                .build(CacheLoader.from(position -> {
-                    byte[] compressedRegion = compressedRegions.get(position);
 
-                    if (compressedRegion == null) {
-                        int x = unpackX(position);
-                        int y = unpackY(position);
-                        return new FlagMap(x * REGION_SIZE, y * REGION_SIZE, (x + 1) * REGION_SIZE - 1, (y + 1) * REGION_SIZE - 1, this.flagCount);
-                    }
+        regionExtents = CollisionMap.getRegionExtents();
+        widthInclusive = regionExtents.getWidth() + 1;
+        final int heightInclusive = regionExtents.getHeight() + 1;
+        regionMaps = new FlagMap[widthInclusive * heightInclusive];
 
-                    try (InputStream in = new GZIPInputStream(new ByteArrayInputStream(compressedRegion))) {
-                        return new FlagMap(Util.readAllBytes(in), this.flagCount);
-                    } catch (IOException e) {
-                        throw new UncheckedIOException(e);
-                    }
-                }));
+        for (Map.Entry<Integer, byte[]> entry : compressedRegions.entrySet()) {
+            final int pos = entry.getKey();
+            final int x = unpackX(pos);
+            final int y = unpackY(pos);
+
+            FlagMap map;
+            try (InputStream in = new GZIPInputStream(new ByteArrayInputStream(entry.getValue()))) {
+                byte[] bytes = Util.readAllBytes(in);
+                map = new FlagMap(bytes, this.flagCount);
+            } catch (IOException e) {
+                throw new UncheckedIOException(e);
+            }
+
+            regionMaps[getIndex(x, y)] = map;
+        }
     }
 
     public boolean get(int x, int y, int z, int flag) {
-        try {
-            return regionMaps.get(packPosition(x / REGION_SIZE, y / REGION_SIZE)).get(x, y, z, flag);
-        } catch (ExecutionException e) {
-            throw new UncheckedExecutionException(e);
+        final int index = getIndex(x / REGION_SIZE, y / REGION_SIZE);
+        if (index < 0 || index >= regionMaps.length || regionMaps[index] == null) {
+            return false;
         }
+
+        return regionMaps[index].get(x, y, z, flag);
+    }
+
+    private int getIndex(int regionX, int regionY) {
+        return (regionX - regionExtents.getMinX()) + (regionY - regionExtents.getMinY()) * widthInclusive;
     }
 
     public static int unpackX(int position) {