Part 2 of performance changes. Last performance PR for now; this brings performance to near real-time levels and any further optimizations I've found either degrade readability or may alter path finding.

I noticed the previous `LoadingCache`'s max size was more than the uncompressed collision map data. It appears the max size was set so high due to empty `FlagMap`s with empty full-size bitsets which kept triggering the weigh constraint to load/unload old regions in the cache. This operation is expensive so I decided to look into if it was even necessary at all.

Changes made:
* Remove the `LoadingCache` and replace it with a minimum-size array similar to `VisitedTiles`.
* Regions with no collision data are null and never loaded; this saves the bulk of memory costs.


The total cost of the map now becomes uncompressed collision data + `FlagMap` member variables + 8 bytes per reference which comes out to... 2,903,256 bytes (2.8MB). The previous maximum cache size in bytes was 20,971,520 (20MB) + all of the compressed collision map data. This change actually ends up storing _less_ long-term data overall.

After this, the only other major optimization that I can find is packing all `WorldPoint`'s into ints in the hotpath but that comes at the expense of readability. I've already demo'd that change and found it reduced total allocations down to ~282MB with avoid wilderness off.

Hopefully after these performance changes, adding more transports/teleports or other features should be more feasible 😄

# Results

**Worst-Case:** GE to top-left corner of the map in the ocean
**Short-Case:** GE to a point west 100 tiles away

| Metric | Before (f596d1d) | After (bef718d) | After (Avoid Wilderness) | Difference |
| --- | --- | --- | --- | --- |
**Worst-Case Time** | 1213.4845ms | 174.7856ms | 151.2286ms | **1038.6989ms** -<br/>**1062.2559ms**
**Short-Case Time** | 28.5309ms | 2.5727ms | 2.1449ms | **25.9582ms** -<br/>**26.386ms**
**Allocations** | 1,286,051,528 bytes | 503,251,296 bytes | 728,338,632 bytes | **782,800,232 bytes** -<br/>**557,712,896 bytes**

## Before Optimization
**Note:** Didn't measure with avoid wilderness on as the difference is not noticeable over noise. Difference in allocation size are the same with/without this PR so see the After results below.

**Worst-Case Time**: 1213.4845ms
**Short-Case Time**: 28.5309ms

**CPU Samples:**
<img width="519" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/5f7047f9-508a-4a2d-8ffb-3bc29a39201f">

**Memory Allocations:**
<img width="601" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/44e826d8-e965-4b31-99fc-548f6b0667aa">

## After Optimization
**Worst-Case Time**: 174.7856ms
**Short-Case Time**: 2.5727ms

Additionally, now that path finding is so quick, there is a clear difference between having avoid wilderness on/off:
**Worst-Case Time** (Avoid Wilderness): 151.2286ms
**Short-Case Time** (Avoid Wilderness): 2.1449ms

Note: for CPU Samples, with/without avoid wilderness didn't make a difference in the call tree as it doesn't make enough samples
**CPU Samples:**
<img width="638" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/9fcc9a3d-3656-49ab-b506-2274c7d5d63a">

**Memory Allocations:**
<img width="542" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/481c9cbd-ab7d-4c7e-b03e-84bc40b31fcd">

**Memory Allocations (Avoid Wilderness):**
<img width="613" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/394e99be-6af0-491f-8a73-2f45315e2079">