@Skretzo 2023-07-16
I have tested this and it works perfectly. I am also seeing comparable performance improvements.

Before I merge this I want to ask for your opinion on the following suggestion I have to maybe improve the flag maps:

Currently, the flags in each flag map always have a length of `width * height * PLANE_COUNT * flagCount = 64 * 64 * 4 * 2`. Most of the upper planes are blocked/unreachable, so maybe turning the `PLANE_COUNT = 4` into a variable can be worthwhile? Of the current total of 1416 regions it is possible to set `planeCount = 1` for 567 regions, `planeCount = 2` for 931 regions and `planeCount = 3` for 1186 regions. This can probably be determined in the `SplitFlagMap` constructor. It will require both flags to be blocked. The image below shows Lumbridge (region 50, 50) which is one of the regions where all 4 planes need to be included (yellow is unblocked, purple is blocked). For Lumbridge it would be theoretically possible to discard z = 4 because that is an unblocked roof which is unreachable, but that would require a more careful approach involving crowdsourced player movement data from the OSRS Wiki.

![image](https://github.com/Skretzo/shortest-path/assets/53493631/6d9789ec-a8cd-4d40-920b-25889002bf0f)