@jocopa3 2023-07-16
Removing empty planes might take the size down a bit more, but uncompressed it's already about the size of a high-res PNG file and compressed is even smaller. I personally wouldn't consider it unless the size was more of an issue (i.e. >10MB).

That said, an easy way of doing it that doesn't require much additional work on either the read or write side is to encode the number of planes in a flagmap using the size of the map; i.e. each plane adds the same amount of data to the map so dividing by the width/height/flag count would give the plane count. Writing could use the same optimization I had for visited tiles; each row is a long and if all rows in a plane are 0 then the plane is empty and can be skipped. One thing to watch out for are any cases where plane 0 and 2 aren't empty but plane 1 is empty; in that situation I'd just write the empty plane to keep the flagmap code simple.

Edge-cases like Lumbridge probably aren't worth optimizing as the potential of having wrong/out-of-date info about unused planes is worse for the user-experience than the 1,024 bytes it saves.