@jocopa3 2023-07-16
One thing I will suggest is you don't need to compress the collision maps if they're going to be zipped up anyway; write them uncompressed then zip the uncompressed files. Makes the file size a bit smaller and saves the cost of uncompressing each file individually at startup.

<img width="424" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/8d4bcba5-e363-4b9b-b0dd-ef128dadcc78">