@Skretzo 2023-07-16
> write them uncompressed then zip the uncompressed files. Makes the file size a bit smaller

Interesting. So both zipping and using GZip is bad. I just continued doing what the original author Runemoro was doing. Would you recommend run-length encoding before zipping? I know that has been used successfully in video encoding. I guess we could also remove the prefixed 16 bytes with minX, minY, maxX and maxY as they can be easily calculated based on the file names (e.g. 50_50) and hardcoding width and height.