@jocopa3 2023-07-16
I wouldn't recommend any compression before zipping, the zip compression will be more clever than most pre-optimizations you could make. That's why the zipped uncompressed files ends up being smaller than the zipped compressed files; compression works better on uncompressed data. It also starts getting into micro-optimization territory when there are much bigger things that could be tackled instead (like the `WorldPoint` packing).