@Skretzo 2023-07-16
I agree that this is micro-optimizations, but I know that the RuneLite devs want to keep hub plugins as small as possible to reduce the bandwidth costs (they have a hard 10 MB hub plugin size limit).

I will go ahead and merge this PR, so thank you very much for your contributions to improve the plugin and its performance! Feel free to also make a PR for the WorldPoint rewrite, as I don't mind the reduced readability too much. I will probably do a release very soon either way. It usually takes about 1 week to be reviewed before it is live in the client.